made document and sub-modules a single module
